♻️ refactor:  constantsという曖昧なファイル名を廃止
Issue #145
